Move LazyWithArgument
It isn’t an extension function, so moved it to the top level package
